Backend/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=1.3.0
numpy>=1.21.0
opencv-python>=4.7.0
Pillow>=9.0.0
supabase
dotenv